requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
requests==2.32.5
urllib3==2.6.0
google-generativeai==0.14.1
google-generativeai-tools==0.1.2
google-auth==2.25.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
flask
python-dotenv